+++ test/cases/exception-execute-source/webpack.config.js
@@ -1,0 +1,36 @@
+const path = require('path');
+const PugPlugin = require('../../../');
+
+const isProduction = true;
+
+module.exports = {
+  mode: isProduction ? 'production' : 'development',
+  devtool: isProduction ? false : 'source-map',
+
+  output: {
+    path: path.join(__dirname, 'public/'),
+    publicPath: '/',
+  },
+
+  entry: {
+    index: './src/index.pug',
+  },
+
+  plugins: [new PugPlugin()],
+
+  module: {
+    rules: [
+      {
+        test: /\.pug$/,
+        use: [
+          {
+            loader: PugPlugin.loader,
+            options: {
+              method: 'render',
+            },
+          },
+        ],
+      },
+    ],
+  },
+};