+++ test/cases/options-enabled/expected/index.js
@@ -1,0 +1,1 @@
+(()=>{var r={545:r=>{r.exports="<h1>Hello World!</h1>"}},o={};!function e(t){var p=o[t];if(void 0!==p)return p.exports;var s=o[t]={exports:{}};return r[t](s,s.exports,e),s.exports}(545)})();