+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change log
 
+## 1.2.1 (2022-01-14)
+- fix: throw an exception if `output.publicPath` is `undefined` or `auto`
+- fix: real source path of required CSS file
+- fix: extract styles and images via require with method `html`
+
 ## 1.2.0 (2022-01-11)
 - added support `webpack serve`
 - added support require of style source directly in pug, e.g.: